mysite/requirements.txt
asgiref==3.7.2
Django==5.0.1
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
Markdown==3.4.1
pillow==10.2.0
psycopg2-binary==2.9.3
sqlparse==0.4.4
typing_extensions==4.9.0
